Stop building libssh
libssh is lgpl, additionally we were only using it for keygeneration
which isn't working for me on android or osx. It only seems to be
working on linux.

At this point, if we want to move away from jsch for generating the
public key, we can build openssh, and use it instead. It seems to be
whatever everyone is using for the keygeneration.