Remove TravisCI
The builds fail on travis, and I do not have a way to figure out why. I
cannot output the logs to stdout, because they are too long and result
in the build being cancelled. I can output them to a file and save that
as an artifact, but Travis only seems to support S3 artifacts, and I
don't want to set that up right now.

So bah.